refactor(PAY-1-A): leaf-extract 5 over-budget LLM adapter files under their line limits
genai-spans -> genai-message-attributes + genai-tool-spans (559/666/318);
sse-accumulator -> sse-frame-splitter (669/103); gateway -> sdk-projection
(705/33); inline-lane index -> inline-lane-output (703/24); query-loop ->
query-tool-success-ledger (688/34). Pure moves; exports preserved;
typecheck + focused unit suites green.
