+++ apps/core/src/adapters/llm/anthropic-claude-agent/runner/query-loop.ts
@@ -2,7 +2,6 @@ import {
   query,
   type EffortLevel,
   type HookInput,
-  type PostToolUseHookInput,
   type ThinkingConfig,
 } from '@anthropic-ai/claude-agent-sdk';
 import { randomUUID } from 'node:crypto';
@@ -80,39 +79,13 @@ import {
 } from '../../../../runner/tool-gate-core.js';
 import { canonicalGantryToolRuleName } from '../../../../shared/gantry-tool-facades.js';
 import { emitJobToolActivity } from './tool-permission-events.js';
+import { recordSuccessfulToolUse } from './query-tool-success-ledger.js';
+
+export { recordSuccessfulToolUse } from './query-tool-success-ledger.js';
 
 interface RunQueryOptions {
   enableIpcFollowups?: boolean;
   persistSdkSession?: boolean;
-}
-
-function toolResponseIsError(response: unknown): boolean {
-  if (Array.isArray(response)) return response.some(toolResponseIsError);
-  if (!response || typeof response !== 'object') return false;
-  const value = response as {
-    is_error?: unknown;
-    isError?: unknown;
-    status?: unknown;
-    error?: unknown;
-    content?: unknown;
-  };
-  return (
-    value.is_error === true ||
-    value.isError === true ||
-    value.status === 'error' ||
-    Boolean(value.error) ||
-    toolResponseIsError(value.content)
-  );
-}
-
-export function recordSuccessfulToolUse(
-  hookInput: Pick<PostToolUseHookInput, 'tool_name' | 'tool_response'>,
-  toolSuccessLedger: RunScopedToolSuccessLedger,
-): void {
-  if (toolResponseIsError(hookInput.tool_response)) return;
-  toolSuccessLedger.recordSuccess(
-    canonicalGantryToolRuleName(hookInput.tool_name),
-  );
 }
 
 function localCliCredentialDirectoriesFromRuntimeAccess(
